Merge sahn_init and sahn_load_file, rename sahn_watch_file
sahn_init now takes an address representing which node represents this instance
in the topology file. sahn_update_topology is for periodic updates of the
topology file , and if the argument can be non-NULL to a different topology
file. A method will be added later for an inotify-based watch of the topology
file, but thread-safety is a prerequisite.